[WIP] Improve benchmarks (#1142)
* start rewriting benchmarks

* azure update

* fix typo

* git use jitted histogram again in fast_kde

* continue benchmark rewrite

* move plot_kde and get_bins

* add kde_utils

* modify benchmarks

* minor changes

* minor modifications to benchmarks

* remove comma

* move get_coords to utils

* move _sturges_formula

* fix pydocstyle

* black changes

* fix benchmarks

* create numerical_utils

* update fast_kde_2d benchmark and changelog

* update lint.sh
