+++ arviz/plots/__init__.py
@@ -20,7 +20,7 @@ from .ppcplot import plot_ppc
 from .rankplot import plot_rank
 from .traceplot import plot_trace
 from .violinplot import plot_violin
-from .plot_utils import _fast_kde, _fast_kde_2d
+from ..numeric_utils import _fast_kde_2d
 
 
 __all__ = [
@@ -35,7 +35,6 @@ __all__ = [
     "plot_hpd",
     "plot_joint",
     "plot_kde",
-    "_fast_kde",
     "_fast_kde_2d",
     "plot_khat",
     "plot_loo_pit",
