+++ arviz/plots/backends/bokeh/posteriorplot.py
@@ -14,9 +14,9 @@ from ...plot_utils import (
     format_sig_figs,
     round_num,
     calculate_point_estimate,
-    get_bins,
 )
 from ....stats import hpd
+from ....numeric_utils import get_bins
 
 
 def plot_posterior(
