+++ arviz/plots/backends/matplotlib/densityplot.py
@@ -8,9 +8,8 @@ from ...plot_utils import (
     make_label,
     _create_axes_grid,
     calculate_point_estimate,
-    _fast_kde,
-    get_bins,
 )
+from ....numeric_utils import _fast_kde, get_bins
 
 
 def plot_density(
