+++ arviz/plots/backends/matplotlib/forestplot.py
@@ -9,8 +9,8 @@ import numpy as np
 from . import backend_kwarg_defaults, backend_show
 from ....stats import hpd
 from ....stats.diagnostics import _ess, _rhat
-from ....stats.stats_utils import histogram
-from ...plot_utils import _scale_fig_size, xarray_var_iter, make_label, get_bins, _fast_kde
+from ....numeric_utils import _fast_kde, histogram, get_bins
+from ...plot_utils import _scale_fig_size, xarray_var_iter, make_label
 from ....utils import conditional_jit
