+++ arviz/plots/elpdplot.py
@@ -6,7 +6,6 @@ from matplotlib.lines import Line2D
 
 from ..data import convert_to_inference_data
 from .plot_utils import (
-    get_coords,
     format_coords_as_labels,
     color_from_dim,
     get_plotting_function,
@@ -14,6 +13,7 @@ from .plot_utils import (
 )
 from ..stats import waic, loo, ELPDData
 from ..rcparams import rcParams
+from ..utils import get_coords
 
 
 def plot_elpd(
