+++ arviz/plots/traceplot.py
@@ -7,13 +7,12 @@ import matplotlib.pyplot as plt
 
 from .plot_utils import (
     get_plotting_function,
-    get_coords,
     xarray_var_iter,
     KwargSpec,
     matplotlib_kwarg_dealiaser,
 )
 from ..data import convert_to_dataset, InferenceData, CoordSpec
-from ..utils import _var_names
+from ..utils import _var_names, get_coords
 from ..rcparams import rcParams
